jobs/influxdb-to-bigquery/requirements.txt
click==8.0.4
pytest==6.0.2
pytest-black==0.3.11
pytest-flake8==1.0.6
influxdb==5.3.1
google-cloud-bigquery==3.5.0
pandas==1.5.1
pyarrow==8.0.0